[FIX] web_editor: duplicate attachments in html field
Before this commit, when an image was cropped in a record on which
multiple users were collaborating, multiple attachments would be created
for the same image (one per user).

Multiple attachments for the same image would also be created if the
user uploaded the same image multiple times in the same record.

This commit makes sure that if:
- an attachment already exists for the same field values, and
- the user about to create an attachment has read rights to it,
the existing attachment is returned instead of creating a new one.

task-3524601