[REF] l10n_at: rework COA and tags ; translate TODO: i18n
This commit reworks the chart of accounts (COA) and tags.
The current tags are no longer needed and are removed.
The main reason for this was the addition of a new module for the
SAF-T export (l10n_at_saft) in the enterprise repository.

In the process all touched files were translated to English.
The translation to German is now handled by the German localization
information. This was done to conform the files to the current guidelines.

The Austrian SAF-T specification defines a COA.
All accounts that are "exported" in the SAF-T document have to
annotated with their corresponding account in the SAF-T COA.

This commit introduces tags to annotate accounts with a code that
should be used for external reports.

For now the new tags will have the following functions:
1. They map the company's chart of accounts (COA) to the SAF-T COA.
2. They are used to assign each account to the correct item / line in
   the Austrian balance sheet or profit and loss report (from enterprise/l10n_at_reports).

To help with point 1 the COA was adjusted such that it can be mapped to the SAF-T COA.

Point 2 above is currently the (only) purpose of the current account tags.
Thus the current account tags are no longer needed and are removed.