[IMP] calendar: always show placeholder in many2many_attendee widget
Currently the Many2ManyTagsAvatarField only shows the placeholder if the field
is empty. In calendar, however, almost never the attendees are empty because
the organizer is always set as an attendee by default. This commit changes the
template to remove the t-if that exists in the Many2ManyTagsAvatarField, so
that the placeholder is always shown.

task-3455051
task-3389317

Part-of: odoo/odoo#138670