[IMP] point_of_sale: self-order (mobile), order page & responsiveness
Before this commit, the module needed certain UI refinements.

- "My orders" page :
The amount and tax were initially separated using justify-align-between,
placing the title on the left and the amount on the right. To align with
Odoo's aesthetics and achieve a ticket-like look, they are now
right-aligned in a table. Additionally, based on feedback, spacing
around the ticket has been removed and the layout has been revamped for
a cleaner appearance.

- Module responsiveness:
Introduced a container for larger screens. This change ensures buttons
and elements don't span the entire width on most of the larger screen
than a mobile. On certain page, the page will dynamically display an
appropriate number of columns depending on the screen's width.
(eg: my orders, product lists,..).

task-3500874