[IMP] product,website_sale:  Add a option to not display the attribute value
Description of the Pull Request:
In the sales module, there is no provision for removing or archiving attribute
values once they have been linked to a product.

it involves the addition of a boolean field at the attribute value level.
Selecting this field will result in the attribute value being concealed from the
following areas: the /shop interface, the product page, and the input fields

Task ID: 3354857