[IMP] base_automation: implement mail and webhook triggers
This commit slightly improves the experience of Base Automation that has been revamped in commit 0a744accc2aaa965d5353e854317895d822ad954.

This commit also implements two major features:
- If a base automation applies to a model that inherits from `mail.thread`, it can react
to mails being sent to or received from a partner outside of the company.
This can be useful to create an activity or some other warning on a record when a customer writes to it.

- A base automation can react to a webhook distant service via the generic route `/web/hook/<some hash>`
This can be useful to change some task's tags when Github sends a relavant payload, or when a payment service sends some data.

Co-authored-by: Damien Bouvy <dbo@odoo.com>

task-id-3450200