[FIX] selection_input: dataseries with full column or full row
Steps to reproduce:
Click on selection input to edit the chart data series.
Click on header "A". The data is now "A:A".
Click on header "B". The data series should be updated with "B:B",
but nothing happens instead.

This issue is created since the boolean waitingForMove is true after
a column is selected. So when user tries to reselect that column,
startMovement method is called followed by an early return hence
startSelection is not called.

This commit fixes the issue by introducing isDefaultSubscription
method in event_stream.ts which returns a boolean based on whether
main subscription is same as default subscription. Based on this boolean,
_selectElement is called to be able to select column/row again while
editing range of dataseries.

Task ID : 3481617