[FIX] account_edi : search checks for identically named product first
If the name of a product on the db contains the name of an imported product in a pdf, the product on the db is used instead of the one on the pdf. Example : If you have a "Displacement" product in your db and you import a pdf with a "Placement" product that does not exist in the db, the product on this line will be set to "Displacement". In such a case the product shouldn't be set and the "Placement" label from the pdf should be kept as is.

To fix this, we first check if there is a name identical to the one in the pdf.

In case we use a product from the db, we also check that the product categories match the one from the pdf to make sure they have comparable units to avoid an error preventing the creation of the invoice.

opw-3470969

co-author Julien Van Roy (JUVR)