[IMP] mass_mailing: allow dynamic filters
ir.filter allows users to do date manipulation to get filters
that are dynamic through time. For example to select all tasks older
than 5 days.

We introduce the ability to do this with mailing filters
which will afford the same flexibility to advanced users to create
similar filters. Which can be very useful for welcoming
campaigns for example.

Additionally removes a double import from
a test file.

task-3223030