[IMP] web_editor: deduce "tel:" links in Link(Tools/Dialog)
When creating or editing a link via the LinkTools or
LinkDialog, http and email URLs can be deduced from incomplete user
input URL (that is, URLs not containing the http(s):// or mailto:
protocol).

This commit extends such URL deduction to telephone numbers, allowing
the user to create a tel:// link without the need to explicitly write
"tel:" or "tel://".

Obs: "tel//:" protocol is preferred over "tel:" for the reasons stated
in [this] commit.

[this]: https://github.com/odoo/odoo/commit/56ce29e71f75a657d4b518d86c1d7084891e898e

task-3516854