[IMP] web_editor: unify http regexes
Before this commit there were two regexes for matching http(s) URLs,
being the only difference between them the presence of a capturing group
for the protocol/scheme. The need for a nearly duplicated version of the
same regex was its use in `_onPaste` as delimiter for String.split(), in
which capturing groups are returned in the resulting array, and the
protocol/scheme was not desired.

This commmit removes the need for two different regexes to match http(s)
URLs, adapting `_onPaste` to work with the presence of the referred
capturing group.