[IMP] payment: improve pending transaction status message
Before this commit, when a transaction enters the pending stage,
customers are shown a warning message that reads "Your payment has been
successfully processed but is waiting for approval".  This seems a bit
confusing as it shows payment successfully but is waiting for approval.

After this commit, we have made the default pending payment message more
understandable with the `Contact Us` link which leads to the chatter of
ongoing transaction order.

task-3323600