[FIX] survey: prevent selecting the same answer multiple times
If multiple questions are conditionally displayed if the user selects an answer
in another multi-choice question, that answer was added multiple times in the
list used to filter mandatory questions that have to be answered when displayed.
This is an issue because the answer was only removed once if the user changes
his choice resulting in a state where a mandatory question that is hidden stays
mandatory even though it is not displayed to the user.

How to reproduce:
- create a new survey with:
  - one page per section
  - 2 sections:
    - Section1:
      - Question1:
        - single line text box
        - multi-choice with one answer
        - 2 answers (A, B)
      - Question2:
        - single line text box
        - mandatory answer
        - conditional display depending on answer B of Question1
      - Question3:
        - conditional display depending on answer B of Question1
    - Section2:
      - Question4:
        - single line text box
- start the survey
- click on answer B then click on answer A

Current behavior:
- user is not able to continue the survey with answer A selected

Expected behavior:
- user should be able to continue the survey

task-

Co-authored-by: Salvo Rapisarda <sdrapisarda@gmail.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>