-- to extract --
[IMP] web_editor: prevent unnecessary elements on text option change

Before this commit, when the users selected the text `[AB]` in a
structure like this `<p>[AB]</p>` and applied a text style (Header 1),
the structure was transformed into `<h1><span>[AB]</span></h1>`.

This commit permits to get rid of the unnecessary span element and
transform the structure into `<h1>[AB]</h1>`.

Before this commit, when the users selected the text `[AB]` in a
structure like this `<p>[AB]</p>` and applied a font size class (small),
the structure was transformed into
`<p><span class="small">[AB]</span></p>`.

This commit permits to get rid of the unnecessary span element and
transform the structure into `<p class="small">[AB]</p>`.

task-1958098