[IMP] website: couple light and dark colors with theme colors
Before this commit, you had a very partial control over the light and
dark colors of Bootstrap: it was linked to the darkest and lightest gray
whose tint and saturation could be controlled (a bit) but not their
lightness. This was ok since light and dark were barely used in
components anyway. Later improvements (and future ones in this PR) will
keep using them more and more, especially in apps layout. We thus need
some more control over them.

This commits does this:

Currently, theme colors are divided into two groups:
o-color-1 o-color-3 o-color-2  +  o-color-4 o-color-5
Only o-color-1 and o-color-3 are saturated colors by default and are
linked to BS components (basically primary and secondary). The others
are not (barely ?) used outside of color combination classes.

Now there will still be two groups but:
o-color-1 o-color-2  +  o-color-3 o-color-4 o-color-5
1 and 2 are still supposed to be the main ones the users should use to
have the color of their company. The 3 others ones will be linked to
light and dark configurations: in color combination classes still but
also for light and dark bootstrap related configurations.

Note: users will still be able to do the same things as before and have
"more than 2 'company' colors": they can still customize their 5 color
combinations classes entirely. So for example, if someone changed
o-color-3 to a bright saturated color to have their second color
combination use that color as background, after upgrade many bootstrap
components will be impacted as it will be linked to the `$light`
bootstrap color. However, they can revert back o-color-3 to a nicer
color to be used as `$light` and force the second color combination
class' background to their old color. That way the visual of their
website will be the same. Some class will still be impacted though
(like o-bg-color-3 which could be added thanks to the secondary tab of
most colorpickers). This is a compromise as this was not "the right way"
to apply a background color anyway, except for icons and text highlights
(in which case users will need to go over their website to force their
old colors if needed) **.

**: Something could be done to help with that in a future commit (maybe
allowing to apply color combinations background on icons for examples).

Also both groups of colors will be named, helping users understand what
they are meant to be used for. Although anything is possible and no
restriction is made on how to pick those 5 colors. Actually, these new
improvements will allow to configure dark themes: just by switching dark
and light colors and/or making the light one a darker color.

task-3097005

Co-authored-by: "stefanorigano (SRI)" <sri@odoo.com>
Co-authored-by: Brieuc-brd <brd@odoo.com>
Co-authored-by: Antoine (anso) <anso@odoo.com>