[ADD] marketing_cards: create dynamic shareable card urls
Word-of-mouth is a great way to get high-quality leads for cheap.
This addon allows users to create templates for partners to share a link
with their a cover image tailored to them on social media.

A good use case is to allow presenters to advertize their track
during an event, each of them can advertize their own track and the
event in a way that's consistent accross all presenters of the event.

This also allows for rewarding partners who do share the event by
giving them any link once they share the event, to a coupon for example.

task-3495279