[MOV] base_automation: merge module with web_studio (EE)
`base_automation` is one of those modules that has existed for a long
time and has provided a lot of customization options in CE.

However, for a few years now the 'home' of customization has been Studio
and its components (model modifications, view and reports modifications,
approvals, etc.). It makes a lot of sense, given the complete rework
that base_automation saw this release cycle, to include it in the
'value' offered by Studio.

The decision to move a module for CE to EE is never taken lightly, and I
have been a supporter of Odoo CE in my years - however in this case I
agree that moving the module is less problematic, as customizations in
CE are usually done mostly by custom modules or server actions (which
remain untouched in CE).

Task-3450200