[IMP] web_editor: add date time option in editor
Specification:

Before this commit we want to add a date and time to it.
With this commit, we have added a new feature that allows users
to add date and time to their content in the editor.
The showDateTime method has been added to the wysiwyg.js file,
which opens a DateTimePicker popover when the user clicks
on the date and time icon. The user can select a date and time from the popover
and the selected date and time will be inserted into the editor.
This commit improves the functionality of the web editor by providing users,
with a convenient way to add date and time to their content.