[FIX] web: resolve draggable kanban records for readonly = "1"
Before this commit:
The kanban records were undraggable only when `readonly="True"` was set, but
draggable for `readonly="1"`.

Reason:
every function for `isReadonly` in  `canMoveRecords` only checks for
`readonly="True"` condition not for `readonly="1"` hence isReadonly becomes
false for `readonly="1"`.

After this commit:
Now this commit enables both `readonly="1"` and `readonly="True"` to control the
draggability of kanban records.

Task-3555209