[IMP] base_automation: move `__action_done` on the transaction
`__action_done` is a mechanism to prevent an automated action
to process itself multiple times while processing itself
e.g. An automated action triggered when writring on a record,
writing on the same record, therefore triggering itself, etc..

The `transaction` is a better place to save this dict
of already processed automations, as triggers could happen
outside the environment where the key `__action_done` was put
in the context, or even in a totally new environment.

It is also a requirement to be able to recompute
stored field with the environment they were triggered with,
instead of the environment processing the flush.
As the key `__action` done is put in the environment of the flush,
and then when using the queued environment to recompute the field,
the `__action_done` is missing, therefore leading to one more
processing of the automation then expected.

e.g.
```
FAIL: BaseAutomationTest.test_050_on_create_or_write_with_create_record
Traceback (most recent call last):
  File "/data/build/odoo/addons/test_base_automation/tests/test_flow.py", line 578, in test_050_on_create_or_write_with_create_record
    self.assertEqual(len(search_result), 1, 'One record on the project model should have been created')
AssertionError: 2 != 1 : One record on the project model should have been created
```