[FIX] web: save the property field on focus out
Before this commit:
When a user enters a property name and clicks outside the popover, it does not
gets saved, the user needs to click on the popover to save it.

Technical reason:
Currently, when the user clicks outside the popover it calls the
onClickAway() method, which triggers the ev.preventDefault() that causes the
issue.

After this commit:
The property name gets saved when the user clicks outside the popover.

Task-3572335