[IMP] account, mail: Send & print mails to followers as well.
Problem
---------
Currently, when using the send & print option on multiple invoices. Only
the customer of the invoice receives the mail not the followers. This
behavior is counter intuitive since when sending one invoice, the
followers of the invoice also get the invoice.

Objective
---------
Send & print multiple invoices should produce the same behavior as when
users sends one mail.

Investigation
-------------
The flow is different when using mass_mailing mechanisms:

 - Select 1 invoice from the view > send & print >
 mail_compose_message._action_send_mail > account_move.message_post >
 mail_thread.notify_thread (this adds a message to the thread and sends
 a mail to the followers)

 - Select 2+ invoices > send & print >
 mail_compose_message._action_send_mail > mail_mail.create >
 mail_mail.send

Solution
---------
Add a context value that specifies if the mail should also be sent to
the followers. If that is the case, add the followers to the recipient
of the mail.

task-3479109