[IMP] core: factorize ir.qweb expresion namespacing into a method
The goal is for `studio` to be able to call that method
to call this method to namespace an expression and then
to be able to use that expression in `safe_eval` passing
dynamic values with `values` in the `locals_dict`,
to mimic the behavior of qweb.

One of the requirement is to be able to default to `None`
variables not assigned.
e.g.:
```py
In [4]: env['ir.qweb']._render(etree.fromstring('<t t-if="full_width">test</t>'))
Out[4]: Markup('')

In [5]: env['ir.qweb']._render(etree.fromstring('<t t-if="not full_width">test</t>'))
Out[5]: Markup('test')

In [6]: from odoo.tools.safe_eval import safe_eval
In [7]: safe_eval("'test' if full_width else ''")
NameError: name 'full_width' is not defined
```