[FIX] hr: traceback when archiving employee
Step to reproduce the bug:
-Go to the employee app
-Go to an employee form view
-Click on the archive button
-Put a description
-Click on the archive button
-> Odoo error

Bug explaination:
The tracking is not implemented for the html field, so when we archive
an employee the tracking cannot work.

Expected behavior:
The employee is archived and the tracking is done without errors.

Bug resolution:
The tracking has been removed from the html field.
The tracking is now replaced by a message in the chatter
in the write method.

Behavior after this commit:
The employee is archived and the tracking is done without errors
with a mesage in the chatter if a value is entered for the
departure_description field has a value.

task-3576659