[IMP] mail: show push to talk unreliable banner on calls
The push-to-talk feature only work within tab focus. This can be
surprising for users that expect this feature to work even when odoo
is not focused. This commit adds a banner that warn the user about
this limitation. If the browser is using a browser compatible with the
ptt extension, also suggest to download it.

task-2664868