[FIX] web: Action "help" content not markup
Issue/Behavior this commit addresses:

Currently, some actions' "help" content are not markuped when passing in
the _loadAction function. This happens because the actionRequest parameter
is neither an action in the actionRegistry or an id or xmlid for an existing
action. It is an object representing an action which is considered as external.
Since it is considered external, it is not trusted and it's "help" content is
not markup-ed by default.

Steps to reproduce the example used for this commit:

1-Install account_accountant
2-Go to the "Aged Payable" report
3-Audit any cell with a null value for a move line (not partner or total)
4-The help content shows in plain text instead of being markup

How the issue/behavior is addressed:

While it is obviously not possible to markup every action's "help" regardless
of their content for security reasons, this commit makes an additionnal check
before returning the untrusted action's content. It checks that the "help" of
the action is the same before and after the action was modified. If it is the
case then it means it is still safe to use and it can be markup-ed safely.

Desired behavior after this commit is merged:

This commit makes it now possible to take an existing action, modify its values
and use it without having to force its "help" content to be markup-ed again but
only if the help of the action has not been modified.