[FIX] pos*: fix tests without demo data
pos*: point_of_sale, pos_online_payment, pos_restaurant, pos_self_order

Prior to this commit, lot of tests were using demo data, which is not
a good way to implement tests as the tests can be runned without demo
data.

This commit fixes the tests by creating the data needed for the tests
at the start of the tests.