[FIX] web: x2many shouldn't have virtualId in domain
Before this commit, if we have x2many fields which contain a virtual
record with the many2many_tags widget, a server error will occur during
a search because the "name_search" will be performed with a domain containing
a virtualId.

This commit will therefore filter the virtualId when generating the domain
used by the "name_search".

X-original-commit: d93446a9bd2048e8b564d44556dac7a21a35896c