[FIX] odoo, web: actions' "help" lose markup state Issue/Behavior this commit addresses:
Currently, some actions' helper content lose their markup state when changed
into an the rpc's response. They are retrieved markuped, they are modified and
still markuped but encoding them makes them lose that status. It happens
because the object carried by a rpc response is unable to hold such data inside
the text and therefore it is lost.

We don't want internal actions to lose their Markup. If they are already
markuped, it makes no sense to have to markup them again on the client side.

Steps to reproduce the examples used for this commit:

IN 16.0 UP TO SAAS-16.4:
1-Create two companies from the same country with their localization package.
2-Create a Tax Unit with both the companies. (main company and VAT nbr have no
impact on the outcome)
3-Open the "Tax Report" report.
4-Select the new Tax Unit in the filters.
5-Click on the "Closing Entry" button in the top left corner.
6-The help content is shown in plain text instead of being markup.

IN 17.0 AND master:
1-Install account_accountant.
2-Go to the "Aged Payable" report.
3-Audit any cell with a null value for a move line (not partner or total)
4-The help content is shown in plain text instead of being markup.

How the issue/behavior is addressed:

While it is obviously not reasonable to markup every action's "help" regardless
of their content for security reasons, this commit adds a check just before the
rpc's response is sent to the client which checks whether the action's helper
is markup or not. If it is, a tag is added to the response so that the client
can notice that it was markuped and lost that status during the transaction
and that it can be markuped safely again.

Desired behavior after this commit is merged:

This commit makes it now possible to take any action in odoo and modify its
values without having to re-markup its helper. This only work for actions
that are handled on the server side so no external action can bypass
the markup security.