[IMP] *: remodel leave behaviour
*:
- resource
- hr_holidays
- appointment_hr
- mrp

This commit gets rid of the `resource.calendar.leaves` model.

__What changed?__
Several models have been added to replace the resource.calendar.leave.
Here are a list of the new models introduced:
- resource.public.leave: handles public leaves that were previously
represented with resource_id=False;
- resource.leave.mixin: an abstract model introducing the base content
of the leaves;
- hr.leave: inherits resource.leave.mixin. A functional change brings
the fact that multi-employees leaves no longer exists;
- appointment.leave: inherits resource.leave.mixin;
- mrp.leave: inherits resource.leave.mixin;
- resource.leave.wizard.mixin: an abstract model for wizards to allow
creating leaves for multiple resources at once;
- hr.holidays.leave.wizard: inherits resource.leave.wizard.mixin;
- appointment.leave.wizard: inherits resource.leave.wizard.mixin;

__What are the reason for this change?__
Three main reasons are why this refactor was introduced:
- Public leaves and individual leaves have very different behaviours;
while public leaves are supposed to be unbound to timezones; public leaves
could relate to multiple calendars or companies so its logic was
handled very differently from other leaves though in the same model;
- Multiple employee leaves are adding complexity in the computation while
also bringing inaccurate information on the record. The duration will be
wrong if the leave is created for two part-time employees;
- Finally, some security issues are involved: as an appointment manager,
it was possible to create leaves for employees, even without hr_holidays,
which might have already brought discrepancies in databases.